+++ Test/cleanup/StringListView.cs
@@ -62,7 +62,7 @@ static class StringListViewCleanup
                 WriteLine($"Duration: {stamp}");
 
                 if (schritt != 20 && schritt != 21 && schritt > 0 && schritt % 2 == 0)
-                    model?.Splice(0, (uint)count);
+                    model?.Splice(0, count);
 
                 if (schritt == 20)
                     scrolledWindow.Ref.Child(Label.New("nil"));
